Use locate theme dir in eject script
